Update sequence window size: client-published to 256, internal to 512
We additionally maintain an internal sequence window, that
is double the size of the sequence window that is published
to the client. This is done because, due to the asynchronous
(= unordered) nature of request processing at the server, it
is possible that in a window of N, assuming sequence numbers
from 1 to N, the server processes and responds to the request
with (N+1) sequence number before the request with sequence
number `1` (that is, the server responds to `N`th request,
then the client sends `N+1`th request, which also gets
processed before the older request with sequence number `1`).
In such a scenario, the request with sequence number `1` will
fall outside the sequence window, and will be dropped. In
order to reduce the chances of such a situation, we maintain
a larger sequence window on the server, while shortening the
client-published sequence window to 256

Change-Id: I9a5f9b0ac7d11e19f1bcff3d6c1e270b2f72ecad