atomic: remove legacy __sync-based implementations
This is effectively the same change as in
https://review.gerrithub.io/c/ffilz/nfs-ganesha/+/1199063, but the
urgency is much less for ntirpc, as there's no proof that the __sync
versions of the functions we have in ntirpc are actually broken. The
problematic functions in nfs-ganesha were the ones that had an `unless`
variant, and those aren't present in ntirpc's copy.

So this change is purely a code cleanup change: the newer functions, the
ones that start with `__atomic`, are over 12 years old and broadly
supported. Instead of maintaining two separate implementations, let's
just maintain a single one.

Signed-off-by: Kyle Lippincott <spectral@google.com>